Update WalkGraph.java
I added comments and changed some variable names.  I did not change the code to "cosolidate recursion" because I have not thought of a way yet to generate the arraylist using a recursive method that only passes in a node.  The hurdle I'm encountering is that I do not know how to prevent duplicates without passing in the arraylist being built.  I could probably just added every node and then remove duplicates afterwards, but I think that approach would be comparably inefficient.